Simplifying configuration model. Shifted responsibility for sensible defaults to configuration builder rather than callers, for consistency.

+++ fork-runner/src/main/java/com/shazam/fork/runner/TestRunFactory.java
@@ -43,7 +43,7 @@ public class TestRunFactory {
                 .withTestPackage(instrumentationInfo.getInstrumentationPackage())
                 .withTestRunner(instrumentationInfo.getTestRunnerClass())
                 .withTestSize(configuration.getTestSize())
-                .withTestOutputTimeout(configuration.getTestOutputTimeout())
+                .withTestOutputTimeout((int) configuration.getTestOutputTimeout())
                 .withCoverageEnabled(configuration.isCoverageEnabled())
                 .build();
